Manually reapply https://github.com/FernCreek/tui.image-editor/commit/0cb3fb49cec0779dcb170016052f0a4b940620a4
Removed logic that changed the origin point of interative text components when the editor is moved into and out
of draw mode. I can't find a good reason why this logic was originally added or why it was removed on our fork.
I'm going to give us the benefit of the doubt and keep this change since the logic seems to accomplish nothing.