Manually reapply changes to cropzone.js from https://github.com/FernCreek/tui.image-editor/commit/12e79429300dfe27857c42b7cbe56c9457dbdc2f
Some of the internals of cropzone were changed significantly so I was unable to manually rebase our changes to it.
The cropzone was changed to be aware of zooming. Each time we got the dimensions of the canvas we are getting the HTML dimensions.
We need to undo the current zoom so we have the dimensions of the canvas in the Fabric Object coordinate system.
When setting Fabric object dimensions and positioning we need to be in the Object coordinate system.
This applies to rendering the initial cropzone background, moving the cropzone, and scaling the cropzone.