Fix jankiness when resizing the cropzone
When resizing the cropzone such that the top or left of the box would move it would jump around as you continuously moved the cursor.
When scaling, Fabric would do some initial scaling before the cropzone specific code was running.
Somehow that was breaking the calculation we would do to determine the new top and/or left of the cropzone.
Resetting Fabric's scaling before doing our calculation seems to consistently fix this.