Prevent crash when moving shape after changing selection
If two shapes were selected and Shift + click was used to deselect one of them the deselection would set _shapeObj on the shape component to null.
This would lead to a crash when trying to move the still selected shape since it would expect the property to have been set.
Given that the shape event handlers have a this of the fabric.Object using _shapeObj in them isn't needed.